+++ rusk/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `abi::public_sender` [#3341]
 - Add `[vm]` config section [#3341]
+- Add simulate transaction API [#1225]
 
 ### Changed
 
@@ -342,6 +343,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1603]: https://github.com/dusk-network/rusk/issues/1603
 [#1371]: https://github.com/dusk-network/rusk/issues/1371
 [#1257]: https://github.com/dusk-network/rusk/pull/1257
+[#1225]: https://github.com/dusk-network/rusk/issues/1225
 [#1219]: https://github.com/dusk-network/rusk/issues/1219
 [#1144]: https://github.com/dusk-network/rusk/issues/1144
 [#970]: https://github.com/dusk-network/rusk/issues/970